pyproject.toml
[project]
name = "extent-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.25.0",
    "json-repair>=0.47.7",
    "pdf2image>=1.17.0",
    "pydantic>=2.11.7",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
]
